refactor(concurrency_managing): hooks spostati in classe dedicata
