+++ interop-qa-tests/src/test/java/it/pagopa/pn/interop/cucumber/steps/SharedStepsContext.java
@@ -1,6 +1,5 @@
 package it.pagopa.pn.interop.cucumber.steps;
 
-import io.cucumber.java.After;
 import io.cucumber.java.Before;
 import io.cucumber.java.Scenario;
 import io.cucumber.spring.ScenarioScope;
@@ -9,8 +8,6 @@ import it.pagopa.interop.authorization.domain.Role;
 import it.pagopa.interop.authorization.service.identity.IdentityService;
 import it.pagopa.interop.authorization.service.utils.PollingService;
 import it.pagopa.interop.common.IHttpExecutor;
-import it.pagopa.pn.interop.cucumber.config.concurrency.ScenariosConcurrencyAuditor;
-import it.pagopa.pn.interop.cucumber.config.concurrency.ScenariosConcurrencyManager;
 import it.pagopa.pn.interop.cucumber.steps.common.*;
 import it.pagopa.pn.interop.cucumber.steps.e_service_template.shared.EServiceTemplateStepContext;
 import it.pagopa.pn.interop.cucumber.steps.notification.model.NotificationCommonContext;
@@ -21,10 +18,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.slf4j.MDC;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.util.StopWatch;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.regex.Matcher;
@@ -59,28 +53,15 @@ public class SharedStepsContext {
     private ProducerKeychainCommonContext producerKeychainCommonContext;
     private TenantCommonContext tenantCommonContext;
 
-    private ScenariosConcurrencyManager concurrencyManager;
-    private ScenariosConcurrencyAuditor concurrencyAuditor;
-
-    // Questa variabile NON è static, quindi è sicura per lo scenario
-    private List<String> acquiredLocks = new ArrayList<>();
-
-    private final StopWatch stopWatch = new StopWatch();
-
     public SharedStepsContext(
             IHttpExecutor httpCallExecutor,
             @Qualifier("interopIdentityService") IdentityService identityService,
             PollingService pollingService,
-            DelayService delayService,
-            ScenariosConcurrencyManager concurrencyManager,
-            ScenariosConcurrencyAuditor concurrencyAuditor) {
+            DelayService delayService) {
         this.httpCallExecutor = httpCallExecutor;
         this.identityService = identityService;
         this.pollingService = pollingService;
         this.delayService = delayService;
-
-        this.concurrencyManager = concurrencyManager;
-        this.concurrencyAuditor = concurrencyAuditor;
     }
 
     @Before
@@ -103,34 +84,6 @@ public class SharedStepsContext {
     public void configLog(Scenario scenario) {
         MDC.clear();
         MDC.put("scenarioId", extractScenarioId(scenario.getName()));
-    }
-
-
-    @Before(order = Integer.MIN_VALUE)
-    public void eventuallyLock(Scenario scenario) {
-        stopWatch.start("Acquisizione Lock");
-
-        this.acquiredLocks = concurrencyManager.acquireLocksFor(scenario);
-        this.concurrencyAuditor.recordStart(scenario.getName());
-
-        stopWatch.stop();
-        System.out.printf("[BEFORE] Scenario: %s - Tempo attesa lock: %d ms%n",
-                scenario.getName(), stopWatch.getLastTaskTimeMillis());
-    }
-
-    @After(order = Integer.MAX_VALUE)
-    public void eventuallyUnlock(Scenario scenario) {
-        try {
-            stopWatch.start("Rilascio Lock");
-            concurrencyAuditor.recordEnd(scenario.getName());
-        } finally {
-            // Il rilascio DEVE essere in un finally per garantire
-            // che i semafori tornino liberi anche se l'auditor fallisce
-            concurrencyManager.releaseLocks(this.acquiredLocks);
-            stopWatch.stop();
-            System.out.printf("[AFTER] Scenario: %s - Tempo rilascio lock: %d ms%n",
-                    scenario.getName(), stopWatch.getLastTaskTimeMillis());
-        }
     }
 
     private static String extractScenarioId(String scenarioName) {
